+++ packages/pose-core/src/types.ts
@@ -48,7 +48,7 @@ export type OnChangeCallbacks = {
 
 export type ActiveActions<A> = Map<string, A>;
 
-export type ActivePoses = Map<string, string>;
+export type ActivePoses = Map<string, string[]>;
 
 export type ChildPosers<V, A, C, P> = Set<Poser<V, A, C, P>>;
 
@@ -62,6 +62,7 @@ export type PoserFactory<V, A, C, P> = (
 
 export interface Poser<V, A, C, P> {
   set: (next: string, props?: Props) => Promise<any>;
+  unset: (toUnset: string, props?: Props) => Promise<any>;
   get: (key?: string) => any;
   has: (key: string) => boolean;
   setProps: (props: Props) => void;
@@ -151,7 +152,7 @@ export type ConvertTransitionDefinition<V, A> = (
 
 export type PoseFactoryConfig<V, A, C, P> = {
   getDefaultProps?: (config: PoserConfig<V>) => Props;
-  defaultTransitions?: Map<string, TransitionFactory<A>>;
+  defaultTransitions?: Map<string, TransitionMap<A> | TransitionFactory<A>>;
   bindOnChange: (
     values: ValueMap<V>,
     onChange: OnChangeCallbacks
@@ -170,4 +171,5 @@ export type PoseFactoryConfig<V, A, C, P> = {
   selectValueToRead: SelectValueToRead<V>;
   extendAPI: ExtendAPI<V, A, C, P>;
   transformPose?: TransformPose<V, A, C, P>;
+  posePriority?: string[];
 };