+++ packages/site/pages/pose/examples.js
@@ -1,4 +1,20 @@
 import ContentTemplate from '~/templates/content/Template';
+import { Centered } from '~/templates/global/grid';
+import styled from 'styled-components';
+import { color } from '~/styles/vars';
+
+const Choose = Centered.withComponent('h2').extend`
+  color: #333;
+  margin: 0 auto;
+  display: block;
+  text-align: center;
+  font-size: 36px;
+  font-weight: bold;
+  border: 2px dashed #333;
+  margin-top: 50px;
+  padding: 40px;
+  border-radius: 20px;
+`;
 
 export default () => (
   <ContentTemplate
@@ -6,5 +22,7 @@ export default () => (
     theme="pose"
     description="Interactive examples for the Pose animation library"
     section="examples"
-  />
+  >
+    <Choose>Choose an example from the menu</Choose>
+  </ContentTemplate>
 );