+++ playground/index.js
@@ -7,8 +7,10 @@ import {
   PoserFLIP,
   PoserPassive,
   PoseDrag,
+  PoseHover,
   PoserManualValues,
   PoseOpacity,
+  PoseFocus,
   PoseFullScreenFlip,
   UnitTypeError
 } from "./plugins/pose";
@@ -20,7 +22,9 @@ import {
   UserPoseInPoseGroup,
   PoseKey,
   PoseKeyFail,
-  Filter
+  Hover,
+  Filter,
+  DragChildren
 } from "./plugins/react-pose";
 
 storiesOf("Plugin: Pose", module)
@@ -29,6 +33,8 @@ storiesOf("Plugin: Pose", module)
   .add("passive values", () => <PoserPassive />)
   .add("manual values", () => <PoserManualValues />)
   .add("drag", () => <PoseDrag />)
+  .add("hover", () => <PoseHover />)
+  .add("focus", () => <PoseFocus />)
   .add("opacity", () => <PoseOpacity />)
   .add("full screen flip", () => <PoseFullScreenFlip />)
   .add("unit type error", () => <UnitTypeError />);
@@ -40,8 +46,10 @@ storiesOf("Plugin: React Pose", module)
   .add("email list", () => <EmailList />)
   .add("user-defined poses within PoseGroup", () => <UserPoseInPoseGroup />)
   .add("pose key test", () => <PoseKey />)
+  .add("hover", () => <Hover />)
   .add("image filter", () => <Filter />)
-  .add("pose key fail test", () => <PoseKeyFail />);
+  .add("pose key fail test", () => <PoseKeyFail />)
+  .add("drag w/ children", () => <DragChildren />);
 
 // import { Decay, DecayModifyTarget, DecayPower, DecayTimeConstant } from './decay';
 // import { Keyframes, KeyframesDuration, KeyframesLoop, KeyframesLinear, KeyframesTimes } from './keyframes';